Suggestions should be working now
The suggestions that are implemented right now is one that only takes
in the group and which week the workouts are for. I am going to update
group to have a team (or location) because I need the postal code for
the weather api. I will be adding more suggestion routes to provide a
variety of options.